adapt skills display and add pages
